refactored Curve w.r.t derivative
refactored

derivative/numeric derivative -> numDiff
TaylorExpansion -> taylorPoly
Curve

Previously, the (optional) exact derivative to a Curve's x and y
functions was set as a property `derivative` of the `x` and `y` objects.
While it makes sense in a sense of the mathematical structure, it's not
very pythonic, has some issues, and is entirely unnecessary. Now instead
the exact derivatives are set as properties `dx` and `dy` of the Curve
object.

Curve was given an `__add__` method, so that complex curves can be
described in terms of simple curves, without needing to write the
equations manually. Will hopefully lead to more intuitive specification
of curves like e.g. "epicycles" with many circular motions. More
importantly, it allows the exact derivative of such objects to be
supplied automatically.